requirements.txt

torch>=2.2.0
transformers>=4.41.0
datasets>=2.19.0
tokenizers>=0.15.2
accelerate>=0.30.0
numpy
tqdm
pyyaml
scikit-learn
matplotlib
